requirements.txt
Pygame
requests
openai
python-dotenv